backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==0.28.0
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
pdfminer.six==20221105